+++ src/apps/update/index.tsx
@@ -1,7 +1,7 @@
 import { UserSettingsLoader } from '../settings/modules/shared/store/storeApi';
 import { getRootContainer } from '../shared';
 import { wrapConsole } from '../shared/ConsoleWrapper';
-import i18n from './i18n';
+import i18n, { loadTranslations } from './i18n';
 import { createRoot } from 'react-dom/client';
 import { I18nextProvider } from 'react-i18next';
 
@@ -16,6 +16,7 @@ async function main() {
   wrapConsole();
 
   let { jsonSettings } = await new UserSettingsLoader().withThemes(false).load();
+  await loadTranslations();
   i18n.changeLanguage(jsonSettings.language);
 
   createRoot(container).render(